refactor: 유효성 검사 테스트는 validationTest.js 파일로 분리
